+++ src/main/java/module/matches/TeamsRatingPanel.java
@@ -23,6 +23,7 @@ import java.util.stream.Stream;
  */
 class TeamsRatingPanel extends LazyImagePanel {
 
+    private static final int ROW_INDEX_INDIRECT_SET_PIECES = 7;
 	private final int RATING_BAR_WIDTH = 18 * UserParameter.instance().fontSize;
 	private final int RATING_BAR_HEIGHT = 4 * UserParameter.instance().fontSize;
 	private final int INSET = UserParameter.instance().fontSize;
@@ -56,7 +57,10 @@ class TeamsRatingPanel extends LazyImagePanel {
         LEFT_DEFENSE(3, "ls.match.ratingsector.leftdefence", Matchdetails::getHomeLeftDef, Matchdetails::getGuestLeftDef, Matchdetails::getGuestRightAtt),
         RIGHT_ATTACK(4, "ls.match.ratingsector.rightattack", Matchdetails::getHomeRightAtt, Matchdetails::getGuestRightAtt, Matchdetails::getGuestLeftDef),
         CENTRAL_ATTACK(5, "ls.match.ratingsector.centralattack", Matchdetails::getHomeMidAtt, Matchdetails::getGuestMidAtt, Matchdetails::getGuestMidDef),
-        LEFT_ATTACK(6, "ls.match.ratingsector.leftattack", Matchdetails::getHomeLeftAtt, Matchdetails::getGuestLeftAtt, Matchdetails::getGuestRightDef);
+        LEFT_ATTACK(6, "ls.match.ratingsector.leftattack", Matchdetails::getHomeLeftAtt, Matchdetails::getGuestLeftAtt, Matchdetails::getGuestRightDef),
+
+        INDIRECT_SET_PIECES_ATTACK(ROW_INDEX_INDIRECT_SET_PIECES, "ls.match.ratingsector.indirect.setpieces.att", Matchdetails::getHomeRatingIndirectSetPiecesAtt, Matchdetails::getGuestRatingIndirectSetPiecesAtt, Matchdetails::getGuestRatingIndirectSetPiecesDef),
+        INDIRECT_SET_PIECES_DEFENSE(ROW_INDEX_INDIRECT_SET_PIECES + 1, "ls.match.ratingsector.indirect.setpieces.def", Matchdetails::getHomeRatingIndirectSetPiecesDef, Matchdetails::getGuestRatingIndirectSetPiecesDef, Matchdetails::getGuestRatingIndirectSetPiecesAtt);
 
         private final int rowIndex;
         private final String labelTranslationKey;
@@ -65,7 +69,11 @@ class TeamsRatingPanel extends LazyImagePanel {
         private final Function<Matchdetails, Integer> versusGuestTeamFunction;
 
         public int getRowNumber() {
-            return getRowIndex() + 1;
+            var ret = getRowIndex() + 1;
+            if (ret > ROW_INDEX_INDIRECT_SET_PIECES){
+                ret++;
+            }
+            return ret;
         }
 
         public int getHomeTeamValue(Matchdetails matchdetails) {
@@ -209,6 +217,11 @@ class TeamsRatingPanel extends LazyImagePanel {
 		m_jgbcBottom.gridx = 2;
 		m_jpBottom.add(m_jlGuestTeamName, m_jgbcBottom);
 
+        var indirectSetPiecesLabel = new JLabel(TranslationFacility.tr("ls.match.ratingsector.indirect.setpieces"));
+        indirectSetPiecesLabel.setFont(generalFont);
+        m_jgbcBottom.insets = new Insets(16, 8, 0, 8);
+        add(indirectSetPiecesLabel, 0,  ROW_INDEX_INDIRECT_SET_PIECES + 1);
+        m_jgbcBottom.insets = new Insets(8, 8, 0, 8);
         Stream.of(RatingSector.values()).forEach(this::addRow);
 		m_jgbcBottom.insets = new Insets(8, 8, 8, 8);
 
@@ -280,9 +293,19 @@ class TeamsRatingPanel extends LazyImagePanel {
 
 	private void setBarsValue(RatingSector ratingSector, Matchdetails matchdetails) {
         var barPar = barPairs.get(ratingSector.getRowIndex());
-        setBarValue(barPar.progressBarHome(), ratingSector.getHomeTeamValue(matchdetails));
-        setBarValue(barPar.progressBarGuest(), ratingSector.getGuestTeamValue(matchdetails));
-	}
+        var homeValue = ratingSector.getHomeTeamValue(matchdetails);
+        var guestValue = ratingSector.getGuestTeamValue(matchdetails);
+        if (homeValue >= 0) {
+            setBarValue(barPar.progressBarHome(), homeValue);
+        } else {
+            resetProgressBar(barPar.progressBarHome());
+        }
+        if (guestValue >= 0) {
+            setBarValue(barPar.progressBarGuest(), guestValue);
+        } else {
+            resetProgressBar(barPar.progressBarGuest());
+        }
+    }
 
     private static void setBarValue(JProgressBar progressBar, int value) {
         double htValue = calcHtValue(value);